producer/requirements.txt
kafka-python==2.0.2
numpy==1.24.3